Add convergence/validation scripts for the paper's Sec. "Convergence and validation"
Implements convergence-validation.md's plan: params.py/metrics.py/plotstyle.py/
refs.py shared infra, quadrature.py (CPU Fibonacci-vs-pseudorandom QMC harness),
and fig_sampling/fig_gridres/fig_deposition/fig_validation/bench.py.

params.py auto-detects GPU memory (and system RAM) to pick a small/medium/large
settings tier, so the same scripts run a smoke-sized sweep on a small GPU and a
much larger one on better hardware (VALIDATION_SCALE env var to override).
run_all.sh runs the full (non-quick) pipeline end to end.

Fixes found and applied while building/validating this:
- src/xigma_i/deposition.py: occupancy arrays were allocated int64, which
  cupy's scatter_add doesn't support on some cupy/CUDA combinations (worked
  on this repo's dev GPU, TypeError'd on another). Switched to int32 (ample
  range for per-cell deposit counts).
- refs.py's streaming table builder had a macroparticle-weight scaling bug
  (chunks were weighted as if each were the whole population, inflating
  total deposited weight by n_particles/chunk_size).
- fig_validation.py's resonance-function width was missing the relativistic
  Doppler stretch (Eq. wR): fixed, and the delta-vs-direct-integration error
  now fits the predicted mu/dwsingle exponent of -1 essentially exactly.
- fig_sampling.py's fitted convergence exponent used a fixed "exclude the
  last 3 points" floor cutoff, which dilutes Fibonacci's reported exponent
  (it floors much sooner than pseudorandom) toward pseudorandom's -- fixed
  to detect each curve's own floor.
- quadrature.py's per-a0-bin interpolation loop was ~30x more expensive than
  necessary (interpolation commutes with summing over a0); replaced with a
  precomputed 3D marginal, verified numerically identical.

Known, expected (not bugs) results surfaced along the way:
- fig_gridres.py's a0-axis resolution scan is exactly flat: the current
  g_sq = 1/(1/s - r_sq) resonance-condition inversion in spectrum_from_table/
  spectrum_kernel_4d does not use a0/ahat at all, so no a0 binning choice can
  change this observable. Worth a look against Eq. wR, which does include
  ahat -- flagged in the script's docstring rather than changed here.
- fig_gridres.py's gamma-resolution scan can turn back upward at the finest
  bin counts if N_p isn't large enough for that resolution (per-cell shot
  noise reintroducing the noise this figure is meant to exclude) -- now
  prints an occupancy diagnostic so this is visible rather than silently
  read as a resolution-effect artifact.

+++ src/xigma_i/deposition.py
@@ -133,7 +133,7 @@ class Table:
     n_particle_samples: int
     total_weight: float
     n_discarded: int
-    occupancy: np.ndarray  # int64, same shape as H: raw (unweighted) deposit counts per cell
+    occupancy: np.ndarray  # int32, same shape as H: raw (unweighted) deposit counts per cell
     gamma_bracket: tuple  # (gamma_lo, gamma_hi) at quantiles (q, 1-q) of the gamma marginal
 
     def save(self, path):
@@ -199,7 +199,12 @@ def deposit_nearest(grid, gamma, theta_x, theta_y, a0, weight, accumulate_dtype=
     H_flat = xp.zeros(int(np.prod(shape)), dtype=accumulate_dtype)
     _scatter_add(xp, H_flat, flat_idx, weight[in_range].astype(accumulate_dtype))
 
-    occ_flat = xp.zeros(int(np.prod(shape)), dtype=xp.int64)
+    # int32, not int64: cupyx.scatter_add (cupy.add.at under the hood) only
+    # supports {int32, float16, float32, float64, uint32, uint64} on some
+    # cupy/CUDA combinations -- int64 works on this codebase's dev GPU but
+    # raised TypeError on another. Per-cell deposit counts fit comfortably
+    # in int32 (max ~2.1e9) for any realistic particle count.
+    occ_flat = xp.zeros(int(np.prod(shape)), dtype=xp.int32)
     _scatter_add(xp, occ_flat, flat_idx, 1)
 
     return H_flat.reshape(shape), occ_flat.reshape(shape), n_discarded
@@ -236,7 +241,7 @@ def deposit_cic(grid, gamma, theta_x, theta_y, a0, weight, accumulate_dtype=np.f
 
     n = gamma.shape[0]
     H_flat = xp.zeros(int(np.prod(shape)), dtype=accumulate_dtype)
-    occ_flat = xp.zeros(int(np.prod(shape)), dtype=xp.int64)
+    occ_flat = xp.zeros(int(np.prod(shape)), dtype=xp.int32)  # see deposit_nearest's comment on int32 vs int64
     n_discarded = 0
 
     corner_bounds = [(i0g, shape[0]), (i0tx, shape[1]), (i0ty, shape[2]), (i0a, shape[3])]
